Break window state out into its own file
